Move subscription implemention alongside execution
...into the same file!
